+++ crates/ruma-client-api/CHANGELOG.md
@@ -6,6 +6,7 @@ Breaking changes:
   as before.
 - Change type of `client_secret` field in `ThirdpartyIdCredentials`
   from `Box<ClientSecret>` to `OwnedClientSecret`
+- Make `id_server` and `id_access_token` in `ThirdpartyIdCredentials` optional
 
 Improvements: